backend/requirements.txt
python-dotenv
openai
langchain_community
langchain
langchain_openai
pymongo
flask
gunicorn